Wizard/ImageSourceSelect: remove Dropdown passthrough component
The on-prem path no longer needs any props forwarded so the
intermediate component adds indirection without value. Render
OnPremImageSourceSelect and HostedImageSourceSelect directly.
